Move broadly-integrated tests to base package
