+++ modules/lily-compiler/src/test/java/io/github/tomboyo/lily/compiler/CompilerSupport.java
@@ -1,4 +1,8 @@
-package io.github.tomboyo.lily.compiler.icg;
+package io.github.tomboyo.lily.compiler;
+
+import static java.nio.file.StandardOpenOption.CREATE;
+import static java.nio.file.StandardOpenOption.TRUNCATE_EXISTING;
+import static java.nio.file.StandardOpenOption.WRITE;
 
 import io.github.tomboyo.lily.compiler.LilyCompiler;
 import io.github.tomboyo.lily.compiler.OasParseException;
